requirement.txt
defusedxml==0.4.1
Django==1.10
djangorestframework==3.4.7
djangorestframework-xml==1.3.0
Pillow==3.3.1
